feat: consultar solicitudes

+++ domain/model/src/main/java/com/co/crediya/requests/model/loanapplication/LoanType.java
@@ -13,7 +13,7 @@ public class LoanType {
   private UUID id;
   private BigDecimal minAmount;
   private BigDecimal maxAmount;
-  private Double interestRate;
+  private BigDecimal interestRate;
   private Boolean autoValidate;
 
   public LoanType(UUID id) {
